+++ src/socket.rs
@@ -1856,104 +1856,6 @@ impl Socket {
     }
 }
 
-#[cfg(target_os = "linux")]
-/// Socket options for the DCCP protocol
-///
-/// for more info see the linux kernel documentation <https://www.kernel.org/doc/html/latest/networking/dccp.html>
-impl Socket {
-    /// Get the service
-    pub fn dccp_service(&self) -> io::Result<u32> {
-        unsafe { getsockopt(self.as_raw(), sys::SOL_DCCP, sys::DCCP_SOCKOPT_SERVICE) }
-    }
-
-    /// Set the service
-    pub fn dccp_set_service(&self, code: u32) -> io::Result<()> {
-        unsafe {
-            setsockopt(
-                self.as_raw(),
-                sys::SOL_DCCP,
-                sys::DCCP_SOCKOPT_SERVICE,
-                code,
-            )
-        }
-    }
-
-    /// Get the current maximum packet size
-    pub fn dccp_cur_mps(&self) -> io::Result<u32> {
-        unsafe { getsockopt(self.as_raw(), sys::SOL_DCCP, sys::DCCP_SOCKOPT_GET_CUR_MPS) }
-    }
-
-    /// Set both the TX and RX CCIDs
-    /// Different TX and RX CCIDs, while in practice allowed, is rarely done and not supported by this library.
-    // todo: allow passing of multiple CCIDs
-    pub fn dccp_set_ccid(&self, ccid: u8) -> io::Result<()> {
-        unsafe { setsockopt(self.as_raw(), sys::SOL_DCCP, sys::DCCP_SOCKOPT_CCID, ccid) }
-    }
-
-    /// Get the current CCID if set.
-    /// The underlying sockopt used here is the DCCP_SOCKOPT_TX_CCID. This means that theoretically only the CCID for TX is returned.
-    /// However, in practice setting different TX and RX CCIDs is rarely used.
-    pub fn dccp_ccid(&self) -> io::Result<u32> {
-        unsafe { getsockopt(self.as_raw(), sys::SOL_DCCP, sys::DCCP_SOCKOPT_TX_CCID) }
-    }
-
-    /// Enables the server to hold the timewait state when closing the connection.
-    /// Only available for listening sockets
-    pub fn dccp_set_server_timewait(&self, hold_timewait: bool) -> io::Result<()> {
-        unsafe {
-            setsockopt(
-                self.as_raw(),
-                sys::SOL_DCCP,
-                sys::DCCP_SOCKOPT_SERVER_TIMEWAIT,
-                hold_timewait as c_int,
-            )
-        }
-    }
-
-    /// Get the value of partial checksum coverage on the sender side
-    pub fn dccp_send_cscov(&self) -> io::Result<u32> {
-        unsafe {
-            getsockopt(self.as_raw(), sys::SOL_DCCP, sys::DCCP_SOCKOPT_SEND_CSCOV)
-        }
-    }
-
-    /// Set the value of partial checksum coverage on the sender side
-    pub fn dccp_set_send_cscov(&self, level: u32) -> io::Result<()> {
-        unsafe {
-            setsockopt(self.as_raw(), sys::SOL_DCCP, sys::DCCP_SOCKOPT_SEND_CSCOV, level)
-        }
-    }
-
-    /// Get the value of partial checksum coverage on the receiver side
-    pub fn dccp_recv_cscov(&self) -> io::Result<u32> {
-        unsafe {
-            getsockopt(self.as_raw(), sys::SOL_DCCP, sys::DCCP_SOCKOPT_RECV_CSCOV)
-        }
-    }
-
-    /// Set the value of partial checksum coverage on the receiver side
-    pub fn dccp_set_recv_cscov(&self, level: u32) -> io::Result<()> {
-        unsafe {
-            setsockopt(self.as_raw(), sys::SOL_DCCP, sys::DCCP_SOCKOPT_RECV_CSCOV, level)
-        }
-    }
-
-    /// Get the maximum length of the output queue
-    pub fn dccp_qpolicy_txqlen(&self) -> io::Result<u32> {
-        unsafe {
-            getsockopt(self.as_raw(), sys::SOL_DCCP, sys::DCCP_SOCKOPT_QPOLICY_TXQLEN)
-        }
-    }
-
-    /// Set the maximum length of the output queue
-    pub fn dccp_set_qpolicy_txqlen(&self, qlen: u32) -> io::Result<()> {
-        unsafe {
-            setsockopt(self.as_raw(), sys::SOL_DCCP, sys::DCCP_SOCKOPT_QPOLICY_TXQLEN, qlen)
-        }
-    }
-    
-}
-
 impl Read for Socket {
     fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
         // Safety: the `recv` implementation promises not to write uninitialised
